Internal reference ID object.
There's a use case in Hydra-land for supporting both internal references
and strings in a single field. It's enabled by AF being backed by
Active-Triples, but in Valkyrie it'll need to be more explicit.

All this means is that for an object to be queryable via joins it needs
to be added as a reference via a Valkyrie::ID rather than just a string
ID.